unit output

+++ Game/BattleLogic/TW_Functions.cpp
@@ -85,7 +85,7 @@ bool IsInTriangle(int x0, int y0, int x1, int y1, int x2, int y2, int x, int y)
 	return true;
 }
 
-bool IsInQuadrangle(int x0, int y0, int x1, int y1, int x2, int y2, int x3, int y3, int x, int y)
+bool IsInQuadrangle(float x0, float y0, float x1, float y1, float x2, float y2, float x3, float y3, float x, float y)
 {
 	if (IsInTriangle(x0, y0, x1, y1, x3, y3, x, y))
 	{
